requirements.txt
fastapi
uvicorn
tensorflow
Pillow
numpy
python-multipart